FEAT: home page, fixed routes, added utils functions
