+++ client/components/Home.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { withRouter, Link } from 'react-router-dom'
+
+
+export default class Home extends Component{
+    render(){
+        return(
+            <div>Home Page</div>
+        )
+    }
+}
