+++ client/components/index.js
@@ -10,7 +10,8 @@ export {default as Linkedlist} from './linkedlist'
 export {default as BinarySearchTree} from './BinarySearchTree'
 export {default as Queue} from './Queue'
 export {default as Stack} from './Stack'
-export {default as AllUserDS} from './AllUserDS'
+export {default as SingleUserDS} from './SingleUserDS'
+export {default as Home} from './Home'
 
 export * from './arrow'
 export * from './DrawNode'
